Refactor default BASE_TAG
This change includes the dash character (`-`) in the default `BASE_TAG`
argument, and all uses of this argument are updated accordingly.  With
the default configuration, there are no changes.

This is necessary in order to support tags that are not prefixed with
the version number.  For example, somebody may use the project to build
images internally using a custom image name and tag as follows,
resulting in an image tagged as `acme/ghc-musl:ghc922`.

    $ earthly --allow-privileged --build-arg BASE_TAG=acme/ghc-musl: +ghc922