Upgrade Alpine packages (#23)
When building an image, the Alpine package index is updated in order to
install various dependencies.  The versions installed depend on the
state of the package index at that time.  One therefore does not
necessarily get the same packages when building `ghc-musl` images at
different times; images are not reproducible in this way.  This commit
adds a command to upgrade any already-installed packages after the
package index is updated.  This ensures that the latest packages are
used, and that the already-installed packages are consistent with the
versions of newly added packages.