Change tag and target syntax (#23)
Previously, each image was tagged like the following:

* `ghc-musl:v24-ghc924`

With this commit, each image is now tagged like the following:

* `ghc-musl:ghc9.2.4-alpine3.16.1-20220730`
* `ghc-musl:ghc9.2.4-alpine3.16.1`
* `ghc-musl:ghc9.2.4`

This allows users to specify image names according to their needs.  An
image tagged with the GHC version, Alpine version, and (UTC) build date
never changes.  An image tagged with just the GHC version and Alpine
version may be updated to include minor/security package releases, but
the fixed Alpine version means that major upgrades should not break
builds.  An image tagged with just the GHC version may be updated to
include new package releases, both major and minor.

The Earthly targets are renamed to match the new tags.  For example,
previous target `ghc924` is now named `ghc9.2.4`.

Note that updating the GHC-only tag should be done with caution, so that
it always points to an image using the latest (supported) Alpine
version.  To facilitate this, the `image` target has a `TAG_GHC`
argument that defaults to `0` so that the GHC-only tag is not saved by
default.  This argument should be set to `1` only when updating the
image using the latest (supported) Alpine version.  The following
command is an example of using this argument when building locally:

```
$ earthly --allow-privileged --build-arg TAG_GHC=1 +ghc9.2.4
```

The `ALPINE_VERSION` variable is still set to the latest (supported)
Alpine version, *not* set to `latest`.  Users who would like to use the
`latest` image are still able to do so by specifying it in a build
argument.